Store form data in PostForm state; clear component state, dispatch action in formClear
