+++ client/js/components/app.js
@@ -1,14 +1,14 @@
 var React = require("react");
 
 var FullPost = require("./full-post");
-var PostForm = require("./post-form");
+var PostFormContainer = require("./post-form-container");
 var PostList = require("./post-list");
 
 var App = function(){
   return(
     <div>
       <FullPost />
-      <PostForm />
+      <PostFormContainer />
       <PostList />
     </div>
   )
