+++ client/js/store.js
@@ -3,7 +3,7 @@ var createStore = redux.createStore;
 var applyMiddleware = redux.applyMiddleware;
 var thunk = require("redux-thunk").default;
 
-var appReducer = require("./reducers/reducer");
-var store = createStore(appReducer, applyMiddleware(thunk));
+var reducer = require("./reducers/reducer");
+var store = createStore(reducer, applyMiddleware(thunk));
 
 module.exports = store;
